refactor: move form and sidebar helpers
